FIX STATO LIKE STRING

+++ CheckStringOnline/src/DAO/CandidaturaDAO.java
@@ -10,7 +10,7 @@ public interface CandidaturaDAO {
 	
 	public abstract ArrayList<Candidatura> vediTutteCandidature() throws ClassNotFoundException ;
 	
-	public abstract ArrayList<Candidatura> resocontoCandidatura(int index) throws ClassNotFoundException;
+	public abstract ArrayList<Candidatura> resocontoCandidatura(long index) throws ClassNotFoundException;
 	
 	public abstract boolean inserisciCandidatura(Candidatura c) throws ClassNotFoundException ;
